moved settings test module
